pyproject.toml
[tool.poetry]
name = "json-tools"
version = "0.1.0"
description = "Tools for processing JSON data"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
packages = [{include = "json_tools"}]

[tool.poetry.dependencies]
python = "^3.9"
python-dateutil = "^2.9.0"

[tool.poetry.group.dev.dependencies]
flake8 = "^7.2.0"
autopep8 = "^2.3.2"

[build-system]
requires = ["poetry-core>=2.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
aw-filter = "json_tools.aw_filter:main"
aw-clean = "json_tools.aw_clean:main"
aw-analyze = "json_tools.aw_analyze:main"
